feat7

+++ Back/src/main/java/com/e106/demolition_king/constructure/repository/ConstructureRepository.java
@@ -3,5 +3,7 @@ package com.e106.demolition_king.constructure.repository;
 import com.e106.demolition_king.constructure.entity.Constructure;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConstructureRepository extends JpaRepository<Constructure, Long> {
+import java.util.List;
+
+public interface ConstructureRepository extends JpaRepository<Constructure, Integer> {
 }
